[IMP] point_of_sale: print preparation ticket from kitchen
In this commit:
-------------------
- Moved common internally used methods to shared logic to avoid duplicating
the same code multiple times

task: 6131467
Enterprise PR: https://github.com/odoo/enterprise/pull/118286
